bot/requirements.txt
aiofiles==24.1.0
aiohttp==3.11.18
aiosqlite==0.21.0
APScheduler==3.11.0
asyncpg==0.30.0
discord.py==2.5.2
Pillow==11.2.1
python-dotenv==0.21.0
discord.py==2.5.2
redis==6.2.0
